modularize main.cpp into core/ and net/ modules (state, conn, registry, socket_util, srt/srtla handlers)
