+++ src/core/state.h
@@ -1,0 +1,36 @@
+/*
+    srtla_rec - SRT transport proxy with link aggregation, forked by IRLToolkit
+    Copyright (C) 2020-2021 BELABOX project
+    Copyright (C) 2024 IRLToolkit Inc.
+    Copyright (C) 2024-2026 OpenIRL
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+// Shared process-wide state. srtla_rec is a single-instance daemon, so the
+// listener socket, downstream SRT address, epoll fd and the group table live
+// here rather than being threaded through every call.
+
+#pragma once
+
+#include <vector>
+#include <sys/socket.h>
+
+#include "core/conn.h"
+
+extern int srtla_sock;
+extern struct sockaddr srt_addr;
+inline constexpr socklen_t addr_len = sizeof(struct sockaddr);
+extern int socket_epoll;
+extern std::vector<srtla_conn_group_ptr> conn_groups;
